ci(tests): add codecov report (#20)
## What this PR does / why we need it:
This PR adds Codecov report generation and uploading to the continuous
integration process. Specifically, it incorporates Codecov to the CI
pipeline, enabling the generation and uploading of coverage reports
after running tests. The addition of this feature enhances code quality
monitoring and facilitates better visibility into test coverage.

## Which issue(s) this PR resolves:
N/A

## Special notes for your reviewer:
Reviewers are encouraged to pay attention to the changes made to the CI
workflow and ensure that the integration with Codecov is correctly
implemented. Additionally, verification of test coverage reporting and
any potential impact on existing CI processes is crucial.

## Additional documentation e.g., enhancement proposals, usage docs,
etc.:
N/A

---------

Signed-off-by: Mateusz Urbanek <mateusz.urbanek.98@gmail.com>