chore(tools): add tools versioning (#21)
## What this PR does / why we need it:
<!--
What code changes are made?
What problem does this PR addresses, or what feature this PR adds?
-->
Added few minor hacks and tweaks to tooling, so we can benefit more from
Renovate for automatic updates of tooling.

## Which issue(s) this PR resolves:
<!--
Usage: `Resolves #<issue number>`, or `Resolves <link to the issue>`.
If PR is about `failing-tests`, please post the related tests in a
comment and do not use `Resolves`
-->
N/A

## Special notes for your reviewer:
<!-- Do you think reviewers should focus on any particular parts of
code? -->
N/A

## Additional documentation e.g., enhancement proposals, usage docs,
etc.:
<!-- This section can be blank if this pull request does not require a
release note. -->
N/A

---------

Signed-off-by: Mateusz Urbanek <mateusz.urbanek.98@gmail.com>